Finishing object references

+++ grails-app/domain/shopify/Product.groovy
@@ -1,7 +1,7 @@
 package shopify
 
 class Product {
-    int id
+    long id
     String title
     static constraints = {
         id generator: 'assigned'
